2D or not 2D...
