+++ priv/static/css/style.css
@@ -1,10 +1,12 @@
-body {
+body,
+svg,
+#graph-container {
   width: 100%;
   height: 100%;
   position: absolute;
   overflow: hidden;
-  padding: 0;
-  margin: 0;
+  margin: 0px;
+  /* background-color: black; */
 }
 
 #mapLabels {
