+++ priv/static/index.html
@@ -44,7 +44,7 @@
     </div>
   </div>
 
-  <div id="mapLabels" hidden=false>
+  <div id="mapLabels" hidden>
     <img src="/images/back.svg" id="backButton" class="round">
     <br>
     <img id="icon" class="round" />
@@ -53,44 +53,9 @@
     <div id="descriptionLabel"></div>
   </div>
 
-  <span id="autoExploreWarning" hidden>AUTO EXPLORING</span>
+  <div id="graph-container"></div>
 
-  <div id="controlsPane">
-    <div id="controls" hidden>
-      <strong>Controls</strong>
-      <br>
-      <br>
-      <strong>[Move]:</strong> WASD
-      <br>
-      <br>
-      <strong>[Up | Down]:</strong> R | F
-      <br>
-      <br>
-      <strong>[Roll]:</strong> Q | E
-      <br>
-      <br>
-      <strong>[Pitch]:</strong> &#8593; | &#8595; or Mouse Drag &#8597;
-      <br>
-      <br>
-      <strong>[Yaw]:</strong> &#8592; | &#8594; or Mouse Drag &#8596;
-      <br>
-      <br>
-      <strong>[Explore]:</strong> Click a Node
-      <br>
-      <br>
-      <strong>[Show current Node]:</strong> G
-      <br>
-      <br>
-      <strong>[View entire Graph]:</strong> Spacebar
-      <br>
-      <br>
-      <strong>[Toggle Autoexplore]:</strong> P
-      <br>
-      <br>
-    </div>
-    <strong>[Show/Hide Controls]:</strong> Esc
-  </div>
-  <script defer src="scripts/bundle.js"></script>
+  <script defer src="bundle.js"></script>
 </body>
 
 </html>
